Split out object listing code from general object utils
Easier to understand separation between the functions if they're in
different modules

No behavioral changes
